requirements.txt
numpy
vkdispatch
matplotlib
pillow